refactor: follow project file naming convention
