+++ src/structures/suggestion-util.ts
